mend

+++ BotRestarter/BotRestarter/CarSorter.cs
@@ -21,15 +21,18 @@ namespace BotRestarter
         void UpdatePicture(int fingerN, string pictureNameFromFinger, string pictureNameFromFinger1)
         {
             string oldPicture = fingerDirectory + fingerN + "\\old" + Path.GetFileName(pictureNameFromFinger1);
-            string newPicture = fingerDirectory + fingerN + "\\" + Path.GetFileName(pictureNameFromFinger1);
+            string newPicture = fingerDirectory + fingerN + "\\" + Path.GetFileName(pictureNameFromFinger1);            
             if (File.Exists(newPicture))
             {
-                if (File.Exists(oldPicture))
+                if (pictureNameFromFinger != newPicture)
                 {
-                    File.Delete(oldPicture);
-                }
-                File.Move(newPicture, oldPicture);
-                NotePad.DoLog("Обновляю " + newPicture);
+                    if (File.Exists(oldPicture))
+                    {
+                        File.Delete(oldPicture);
+                    }
+                    File.Move(newPicture, oldPicture);
+                    NotePad.DoLog("Обновляю " + newPicture);
+                }                
             }
             File.Move(pictureNameFromFinger, newPicture);
         }
@@ -39,6 +42,7 @@ namespace BotRestarter
             NotePad.ClearLog();
             string[] picturesFromFinger1 = readFilesFromDirectory(finger1Path);
             NotePad.DoLog("В Finger1 " + picturesFromFinger1.Length + " файлов");
+            
             for (int fingerN = 2; fingerN < 6; fingerN++)
             {
                 string[] picturesFromFinger = readFilesFromDirectory(fingerDirectory + fingerN);
@@ -51,7 +55,7 @@ namespace BotRestarter
                         Bitmap pictureFromFinger = new Bitmap(pictureNameFromFinger);
                         foreach(string pictureNameFromFinger1 in picturesFromFinger1)
                         {
-                            if (pictureNameFromFinger1.Contains("jpg"))
+                            if (pictureNameFromFinger1.Contains("jpg") && !pictureNameFromFinger1.Contains("old"))
                             {
                                 Bitmap pictureFromFinger1 = new Bitmap(pictureNameFromFinger1);
                                 int shadesdifs = 0;
@@ -92,6 +96,53 @@ namespace BotRestarter
                     }                    
                 }                
             }
+            
+            //update
+            for(int number = 1; number < 1500; number++)
+            {
+                string currentPictureName = finger1Path + "\\" + number + ".jpg";
+                if (File.Exists(currentPictureName))
+                {
+                    Bitmap currentPicture = new Bitmap(currentPictureName);
+                    bool found = false;
+
+                    for (int fingerN = 2; fingerN < 6; fingerN++)
+                    {
+                        if (!found)
+                        {
+                            string[] picturesFromFinger = readFilesFromDirectory(fingerDirectory + fingerN);
+                            foreach (string pictureNameFromFinger in picturesFromFinger)
+                            {
+                                if (!pictureNameFromFinger.Contains("unsorted") && !pictureNameFromFinger.Contains("old"))
+                                {
+                                    Bitmap pictureFromFinger = new Bitmap(pictureNameFromFinger);
+                                    int shadesdifs = 0;
+                                    for (int x = 0; x < 50; x++)
+                                    {
+                                        for (int y = 0; y < 50; y++)
+                                        {
+                                            var colorValue0 = currentPicture.GetPixel(x + x0, y + y0);
+                                            var colorValue1 = pictureFromFinger.GetPixel(x + x0, y + y0);
+                                            shadesdifs += (Math.Abs((int)colorValue0.R - (int)colorValue1.R) +
+                                                Math.Abs((int)colorValue0.G - (int)colorValue1.G) +
+                                                Math.Abs((int)colorValue0.B - (int)colorValue1.B));
+                                        }
+                                    }
+                                    pictureFromFinger.Dispose();
+                                    if (shadesdifs < 60000)
+                                    {
+                                        found = true;
+                                        currentPicture.Dispose();
+                                        UpdatePicture(1, currentPictureName, pictureNameFromFinger);
+                                        break;
+                                    }
+                                }
+                            }
+                        }
+                    }
+                    currentPicture.Dispose();
+                }
+            }       
 
             NotePad.DoLogWithTime("найдено машин: " + foundcars);
             NotePad.DoLogWithTime("вероятных совпадений " + predictcars);
